Fix priority ignore when 0
Fixed the issue that priority is still output when priority is 0. The operator `!==` will also compare the types. `$priority` is of type float and `0` is implicitly int. Therefore, no matter what the `$priority` value is, it will return `TRUE` because of the different types. Therefore, need to change `0` to `0.0`, or more correctly, change `!==` to `!=` which have implicit type conversion.